Feat/unit test

+++ db/migration/000001_init_schema.up.sql
@@ -26,7 +26,7 @@ CREATE TABLE "coffee" (
 );
 
 CREATE TABLE "machine" (
-  "id" integer PRIMARY KEY,
+  "id" bigserial PRIMARY KEY,
   "sector" varchar NOT NULL,
   "company" varchar NOT NULL,
   "coffee_id" bigint NOT NULL,
